refactor: create read-package-hook package
