+++ .changeset/curvy-socks-press.md
@@ -1,0 +1,5 @@
+---
+"@pnpm/hooks.read-package-hook": major
+---
+
+First release.
