+++ packages/core/package.json
@@ -26,6 +26,7 @@
     "@pnpm/graph-sequencer": "1.0.0",
     "@pnpm/headless": "workspace:*",
     "@pnpm/hoist": "workspace:*",
+    "@pnpm/hooks.read-package-hook": "workspace:*",
     "@pnpm/lifecycle": "workspace:*",
     "@pnpm/link-bins": "workspace:*",
     "@pnpm/lockfile-file": "workspace:*",
@@ -39,7 +40,6 @@
     "@pnpm/normalize-registries": "workspace:*",
     "@pnpm/npm-package-arg": "^1.0.0",
     "@pnpm/package-requester": "workspace:*",
-    "@pnpm/parse-overrides": "workspace:*",
     "@pnpm/parse-wanted-dependency": "workspace:*",
     "@pnpm/prune-lockfile": "workspace:*",
     "@pnpm/read-modules-dir": "workspace:*",
@@ -52,7 +52,6 @@
     "@pnpm/symlink-dependency": "workspace:*",
     "@pnpm/types": "workspace:*",
     "@pnpm/which-version-is-pinned": "workspace:*",
-    "@yarnpkg/extensions": "1.2.0-rc.1",
     "@zkochan/rimraf": "^2.1.2",
     "dependency-path": "workspace:*",
     "is-inner-link": "^4.0.0",
