+++ packages/hooks.read-package-hook/README.md
@@ -1,0 +1,13 @@
+# @pnpm/hooks.read-package-hook
+
+> Creates the default package reader hook used by pnpm
+
+## Installation
+
+```sh
+pnpm add @pnpm/hooks.read-package-hook
+```
+
+## License
+
+MIT
