+++ packages/hooks.read-package-hook/jest.config.js
@@ -1,0 +1,3 @@
+const config = require('../../jest.config.js');
+
+module.exports = Object.assign({}, config, {});
