+++ packages/hooks.read-package-hook/src/createReadPackageHook.ts
