+++ packages/hooks.read-package-hook/src/index.ts
@@ -1,0 +1,1 @@
+export { createReadPackageHook } from './createReadPackageHook'
