+++ packages/hooks.read-package-hook/test/createPeerDependencyPatcher.test.ts
@@ -1,4 +1,4 @@
-import createPeerDependencyPatcher from '../../lib/install/createPeerDependencyPatcher'
+import createPeerDependencyPatcher from '../lib/createPeerDependencyPatcher'
 
 test('createPeerDependencyPatcher() ignores missing', () => {
   const patcher = createPeerDependencyPatcher({
