+++ packages/hooks.read-package-hook/tsconfig.json
@@ -1,0 +1,25 @@
+{
+  "extends": "@pnpm/tsconfig",
+  "compilerOptions": {
+    "outDir": "lib",
+    "rootDir": "src"
+  },
+  "include": [
+    "src/**/*.ts",
+    "../../typings/**/*.d.ts"
+  ],
+  "references": [
+    {
+      "path": "../matcher"
+    },
+    {
+      "path": "../parse-overrides"
+    },
+    {
+      "path": "../parse-wanted-dependency"
+    },
+    {
+      "path": "../types"
+    }
+  ]
+}
